feat: replace Debug with admin-only Admin panel
- Create AdminConfig with admin email (jamiebright1@gmail.com) and
  isCurrentUserAdmin helper backed by Supabase auth
- Rename Debug → Admin in home menu, only visible to admin email
- New AdminScreen with Supabase-backed admin tools:
  - Gift gold / levels / flights to any user by username
  - Change usernames (moderation for offensive names)
  - Gift cosmetics
  - Quick self-actions (unlimited gold, XP, flights)
  - Game log viewer (preserved from debug screen)
- All gift dialogs include validation and error feedback
- Username uniqueness check before signup in auth_service
- Remove TestAccounts (fake accounts no longer needed)
- Replace isDebugMode with computed isAdmin property
- Username change catches Postgres unique constraint violations
